Improve handling of active instances
