+++ client/src/containers/Toolbar/Toolbar.tsx
@@ -1,43 +1,29 @@
 import { Link, useLocation } from "react-router-dom";
 import { ReactComponent as SettingsSVG } from "../../assets/icons/settings.svg";
 import { ReactComponent as EditSquareSVG } from "../../assets/icons/edit-square.svg";
-import { ReactComponent as DoneAllSVG } from "../../assets/icons/done-all.svg";
-import { ReactComponent as CloseSVG } from "../../assets/icons/close.svg";
+import { ReactComponent as TimerSVG } from "../../assets/icons/timer.svg";
 import Menu from "./ToolbarElements/Menu";
 import Outlines from "./ToolbarElements/Outlines";
 import SlideEditTools from "./ToolbarElements/SlideEditTools";
 import Undo from "./ToolbarElements/Undo";
 import UserSection from "./ToolbarElements/UserSection";
-import { useSelector, useDispatch } from "../../hooks";
+import QuickLinkSelection from "./ToolbarElements/QuickLinkSelection";
+import { useSelector } from "../../hooks";
 import { forwardRef, useContext, useEffect, useMemo, useState } from "react";
 import Button from "../../components/Button/Button";
-import { ReactComponent as TimerSVG } from "../../assets/icons/timer.svg";
 import TimerControls from "../../components/TimerControls/TimerControls";
 import { ControllerInfoContext } from "../../context/controllerInfo";
-import {
-  setSelectedQuickLink,
-  setSelectedQuickLinkPresentation,
-} from "../../store/preferencesSlice";
-import { OverlayInfo, Presentation } from "../../types";
+import cn from "classnames";
 
 type sections = "settings" | "slide-tools" | "timer-manager";
 
 const Toolbar = forwardRef<HTMLDivElement, { className: string }>(
   ({ className }, ref) => {
     const location = useLocation();
-    const dispatch = useDispatch();
-
-    const { isEditMode, type, selectedSlide, slides, name, timerInfo } =
-      useSelector((state) => state.undoable.present.item);
-
-    const { selectedQuickLink } = useSelector(
-      (state) => state.undoable.present.preferences
+    const { isEditMode, type } = useSelector(
+      (state) => state.undoable.present.item
     );
-
-    const overlayInfo = useSelector((state) => state.undoable.present.overlays);
-
     const [section, setSection] = useState<sections>("settings");
-
     const { isMobile } = useContext(ControllerInfoContext) || {};
 
     const onItemPage = useMemo(
@@ -78,9 +64,11 @@ const Toolbar = forwardRef<HTMLDivElement, { className: string }>(
               disabled={!onItemPage}
               svg={EditSquareSVG}
               onClick={() => setSection("slide-tools")}
-              className={`text-xs rounded-none ${
-                section === "slide-tools" && "bg-gray-800"
-              }`}
+              className={cn(
+                "text-xs rounded-none",
+                section === "slide-tools" && "bg-gray-800",
+                !onItemPage && "hidden"
+              )}
             >
               Slide Tools
             </Button>
@@ -89,22 +77,25 @@ const Toolbar = forwardRef<HTMLDivElement, { className: string }>(
               svg={TimerSVG}
               disabled={!onItemPage || type !== "timer"}
               onClick={() => setSection("timer-manager")}
-              className={`text-xs rounded-none ${
-                section === "timer-manager" && "bg-gray-800"
-              }`}
+              className={cn(
+                "text-xs rounded-none",
+                section === "timer-manager" && "bg-gray-800",
+                !onItemPage && "hidden"
+              )}
             >
               Timer Manager
             </Button>
           </div>
 
           <div
-            className={`px-2 py-1 flex gap-1 items-center flex-1 ${
-              isEditMode ? "hidden" : ""
-            }`}
+            className={cn(
+              "px-2 py-1 flex gap-1 items-center flex-1",
+              isEditMode && "hidden"
+            )}
           >
-            <Outlines className={`${section !== "settings" && "hidden"}`} />
+            <Outlines className={cn(section !== "settings" && "hidden")} />
             <Button
-              className={`${section !== "settings" && "hidden"}`}
+              className={cn(section !== "settings" && "hidden")}
               variant="none"
               svg={SettingsSVG}
             >
@@ -112,147 +103,14 @@ const Toolbar = forwardRef<HTMLDivElement, { className: string }>(
                 Preferences
               </Link>
             </Button>
-            {selectedQuickLink && selectedQuickLink.linkType === "slide" && (
-              <section className="flex justify-center rounded-l-md rounded-r-md mr-2">
-                <Button
-                  variant="cta"
-                  className="justify-center rounded-r-none"
-                  disabled={!slides?.[selectedSlide]}
-                  svg={DoneAllSVG}
-                  onClick={() => {
-                    let title = "";
-                    let text = "";
-
-                    if (type === "bible") {
-                      title =
-                        selectedSlide > 0
-                          ? slides[selectedSlide].boxes[2]?.words || ""
-                          : slides[selectedSlide].boxes[1]?.words || "";
-                      text =
-                        selectedSlide > 0
-                          ? slides[selectedSlide].boxes[1]?.words || ""
-                          : "";
-                    }
-
-                    dispatch(
-                      setSelectedQuickLinkPresentation({
-                        name,
-                        slide: slides[selectedSlide],
-                        type: "slide",
-                        timerId: timerInfo?.id,
-                        bibleDisplayInfo: {
-                          title,
-                          text,
-                        },
-                      })
-                    );
-                  }}
-                >
-                  {isMobile ? "Select" : "Select Slide"}
-                </Button>
-                <Button
-                  variant="secondary"
-                  className="justify-center rounded-l-none"
-                  svg={CloseSVG}
-                  onClick={() => {
-                    dispatch(setSelectedQuickLink(""));
-                  }}
-                >
-                  {isMobile ? "" : "Cancel Selection"}
-                </Button>
-              </section>
-            )}
-            {selectedQuickLink && selectedQuickLink.linkType === "overlay" && (
-              <section className="flex justify-center rounded-l-md rounded-r-md">
-                <Button
-                  variant="cta"
-                  className="justify-center rounded-r-none"
-                  disabled={!overlayInfo?.id}
-                  svg={DoneAllSVG}
-                  onClick={() => {
-                    let presentationInfo: Presentation = {
-                      name: overlayInfo.name || overlayInfo.description || "",
-                      slide: null,
-                      type: "overlay",
-                    };
-
-                    const info: OverlayInfo = {
-                      id: overlayInfo.id,
-                      type: overlayInfo.type,
-                      duration: overlayInfo.duration,
-                    };
-
-                    if (overlayInfo.type === "participant") {
-                      presentationInfo = {
-                        ...presentationInfo,
-                        participantOverlayInfo: {
-                          ...info,
-                          name: overlayInfo.name,
-                          event: overlayInfo.event,
-                          title: overlayInfo.title,
-                        },
-                      };
-                    }
-
-                    if (overlayInfo.type === "stick-to-bottom") {
-                      presentationInfo = {
-                        ...presentationInfo,
-                        stbOverlayInfo: {
-                          ...info,
-                          subHeading: overlayInfo.subHeading,
-                          title: overlayInfo.title,
-                        },
-                      };
-                    }
-
-                    if (overlayInfo.type === "qr-code") {
-                      presentationInfo = {
-                        ...presentationInfo,
-                        qrCodeOverlayInfo: {
-                          ...info,
-                          url: overlayInfo.url,
-                          description: overlayInfo.description,
-                          color: overlayInfo.color,
-                        },
-                      };
-                    }
-
-                    if (overlayInfo.type === "image") {
-                      presentationInfo = {
-                        ...presentationInfo,
-                        imageOverlayInfo: {
-                          ...info,
-                          imageUrl: overlayInfo.imageUrl,
-                        },
-                      };
-                    }
-
-                    dispatch(
-                      setSelectedQuickLinkPresentation(presentationInfo)
-                    );
-                  }}
-                >
-                  {isMobile ? "Select" : "Select Overlay"}
-                </Button>
-                <Button
-                  variant="secondary"
-                  className="justify-center rounded-l-none"
-                  svg={CloseSVG}
-                  onClick={() => {
-                    dispatch(setSelectedQuickLink(""));
-                  }}
-                >
-                  {isMobile ? "" : "Cancel Selection"}
-                </Button>
-              </section>
-            )}
+            <QuickLinkSelection isMobile={isMobile} />
             <SlideEditTools
-              className={`${section !== "slide-tools" && "hidden"}`}
+              className={cn(section !== "slide-tools" && "hidden")}
             />
             <TimerControls
-              className={`${
+              className={cn(
                 (section !== "timer-manager" || type !== "timer") && "hidden"
-              }`}
+              )}
             />
           </div>
         </div>
