+++ client/src/store/store.ts
@@ -843,7 +843,9 @@ const rootReducer: Reducer = (state: RootState, action: Action) => {
 const store = configureStore({
   reducer: rootReducer,
   middleware: (getDefaultMiddleware) =>
-    getDefaultMiddleware().prepend(listenerMiddleware.middleware),
+    getDefaultMiddleware({
+      serializableCheck: false,
+    }).prepend(listenerMiddleware.middleware),
 });
 
 export default store;
